all: replace size() macro with SIZEOF() macro
Attempting to build with gcc for C++17 emits many errors because the size()
macro conflicts with the in-built C++17 std::size(). Rename the size() macro
to SIZEOF() since it effectively a cast version of the standard sizeof()
function, and is unique enough to enable it to be easily changed later if
required.