all: convert from byte type to stdint uint8_t type
The byte type is defined in kauai/src/util.h as a typedef to unsigned char but if
that line is commented out then the build still succeeds. This is because if
<windows.h> is included then eventually a header called rpcndr.h is included which
also defines byte as a typedef to unsigned char.

Rather than allow this confusion to continue, go ahead and replace all instances
of byte within the 3DMMEx source with uint8_t. This should also help building
on platforms with different type sizes from x86 Windows.