cmake: add GenerateChunkTags.cmake script
The chunk tags in use by kauai are used both as int32_t constants within
the application as well as strings when compiling the chunk files.

Soon we will need to generate separate int32_t and string versions of the
chunk tags: instead of doing this by hand, add GenerateChunkTags.cmake as
a cmake script that can generate the existing chunk tag definitions from
an input file.