inc/tagman.h: move pcrf field to end of TAG structure
The TAG structure cannot be used on non-32 bit platforms in its existing
form because the embedded pcrf pointer is limited to 4 bytes.

Now that all the places where TAG is being serialised to disk have been
updated to use TAGF, we can move the pcrf pointer to the end of the TAG
structure (renaming the previous pcrf field to a dummy _pcrf field) to
allow its use on 64-bit platforms.

Increasing the TAG structure size causes several of the
VERIFY_STRUCT_SIZE() checks to fail, but these can now be removed since
they have been replaced by their on-disk equivalents.