src/studio/stdiobrw.cpp: use TAGF for new scene browser selection
The studio browser currently uses a TAG to pass information about the scene
selection back to the studio, however if the size of TAG increases then it
will no longer fit within the rglw parameter array.

Since the pcrf field isn't used, we can simply use TAGF instead of TAG for
passing the scene selection and avoid triggering an Assert() if the size of
TAG increases.