src/studio/portf.cpp: replace GetWindowLong() with GetWindowLongPtr()
The portfolio makes use of GetWindowLong() with GWL_USERDATA for
retrieving user-defined Window data. Unfortunately this API function
cannot return a pointer on 64-bit Windows platforms because the final LONG
parameter is still restricted to 32 bits causing the pointer to become
truncated.

Replace GetWindowLong() with the newer GetWindowLongPtr() API function
which is guaranteed to work correctly on both 32-bit and 64-bit Windows
platforms.