src/CMakeLists.txt: fix 32-bit dependency during 64-bit link
Commit 3dc5b43c "Enable visual styles for common controls" introduced
a manifest dependency on the common control library to a specific
version to enable visual styles. However the processorArchitecture was
hardcoded to x86 which introduces a 32-bit dependency to the final
3dmovie.exe generated on 64-bit Windows causing it to fail with a
0xc000007b error upon launch.

Update the previous target_link_options() to improve the formatting
and change processorArchitecture to use ${buildKitTargetArch} so that
32-bit and 64-bit Windows platforms will link in the correct
architecture library.