src/studio/portf.cpp: replace SetWindowLong() with SetWindowLongPtr()
The portfolio makes use of SetWindowLong() with GWL_USERDATA for storing
user-defined Window data. Unfortunately this API function cannot store a
pointer on 64-bit Windows platforms because the final LONG parameter
is still restricted to 32 bits causing the pointer to become truncated.

Replace SetWindowLong() with the newer SetWindowLongPtr() API function
which is guaranteed to work correctly on both 32-bit and 64-bit Windows
platforms.