kauai/src/menuwin.cpp: use uintptr_t instead of int32_t for lw params
The menu interface attempts to pass a pointer (HWND) as a lw parameter
which will not work on non-32 bit platforms. Update the menu interface
to accept uintptr_t instead of int32_t and update all callers where
appropriate.

Incorporated into this is an update to the cidChooseWnd command which
currently uses rglw[0] to pass the target window HWND, which will fail
because rglw is an array of int32_t items whilst HWND is a pointer
which can be more than 4 bytes on non-32 bit platforms.

Resolve the issue by storing and retrieving the target window HWND
using a *(uintptr_t *)rglw access such that the HWND will be stored in
rglw[0] for 32-bit systems and rglw[0] and rglw[1] for non-32 bit
platforms.