src/engine/scene.cpp: fix scene start event pointer conversion
DeserializeStartEVs() converts the on-disk TAGF structure into a 32-bit
tag ready for use in PscenRead(). However during part of the scene
reading process, the 32-bit tags are overwritten with pointers to the
relevant object instead.

As this is done with the Put() method the size of the stored item is
not updated causing the pointer to be truncated on 64-bit Windows.
Change the logic in PscenRead() to use FPut() which also allows for the
size of the stored value to be increased as required on 64-bit Windows.