Refactor APP::_FFindMsKidsDir to change search order
Previously, 3DMM would read the install directory from the registry.
If not found, it would search the application's directory and the
current working directory for the Microsoft Kids or MSKIDS
directories. On systems with 3DMM installed, this would cause 3DMMEx
to load Chunky files from the installed copy of 3DMM instead of the
3DMMEx directory.

This commit changes _FFindMsKidsDir to search the application's
directory first, followed by the install directory from the
registry (if present) and the current working directory.