src/studio/utest.cpp: move DWORD declarations inside #ifdef WIN32
The existing #ifdef WIN32 is in the correct place, however it fails to compile
on non-Windows platforms because the DWORD declarations are outside of it.

Also separate the initialisation from the declaration as otherwise MSVC
generates an error because it doesn't like that the static initialiser can occur
after the goto.